# swagger.yaml
openapi: 3.0.0
info:
  title: Reality Puchýř API
  description: API documentation for Reality Puchýř real estate platform
  version: 1.0.0
  contact:
    name: Reality Puchýř Support
    email: support@realitypuchyr.cz
servers:
  - url: /api
    description: API base URL
...
components:
  schemas:
...
    Blog:
      type: object
      required:
        - name
        - slug
        - content
        - language
      properties:
        id:
          type: integer
          description: Blog post ID
        name:
          type: string
          description: Blog post title
        slug:
          type: string
          description: URL-friendly slug
        content:
          type: string
          description: Blog post content
        tags:
          type: array
          items:
            type: string
          description: Blog post tags
        date:
          type: string
          format: date-time
          description: Publication date
        pictures:
          type: array
          items:
            type: string
          description: Array of image URLs
        metaTitle:
          type: string
          description: SEO meta title
        metaDescription:
          type: string
          description: SEO meta description
        keywords:
          type: string
          description: SEO keywords
        language:
          type: string
          enum: [cs, en, de]
          default: cs
          description: Content language
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    BlogTranslation:
      type: object
      required:
        - blogId
        - language
        - name
        - slug
        - content
      properties:
        id:
          type: integer
          description: Translation ID
        blogId:
          type: integer
          description: ID of the blog
        language:
          type: string
          enum: [cs, en, de, ru, ua, vn, es, fr, it]
          description: Language code
        name:
          type: string
          description: Translated blog title
        slug:
          type: string
          description: Translated blog slug
        content:
          type: string
          description: Translated blog content
        tags:
          type: array
          items:
            type: string
          description: Translated blog tags
        metaTitle:
          type: string
          description: Translated SEO meta title
        metaDescription:
          type: string
          description: Translated SEO meta description
        keywords:
          type: string
          description: Translated SEO keywords
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Error message
              details:
                type: string
                description: Additional error details
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
security:
  - ApiKeyAuth: []
paths:
...
  /blogs:
    get:
      tags:
        - Blog
      summary: Get all blogs
      description: Returns paginated blog posts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Items per page
        - name: truncate
          in: query
          schema:
            type: integer
            default: 0
          description: Number of characters to truncate content
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Blog'
        '500':
          $ref: '#/components/responses/Error'
  /blogs/{slug}:
    get:
      tags:
        - Blog
      summary: Get blog by slug
      description: Returns a single blog post by its slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blog'
        '404':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
  /blogs/{id}/translate:
    post:
      tags:
        - Blog
      summary: Translate blog to target language
      description: Translates a blog post to the specified target language using DeepL API
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Blog ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - targetLanguage
              properties:
                targetLanguage:
                  type: string
                  enum: [en, cs, de, ru, ua, vn, es, fr, it]
                  description: Target language code
                sourceLanguage:
                  type: string
                  enum: [en, cs, de, ru, ua, vn, es, fr, it]
                  description: Source language code (optional, defaults to Czech)
      responses:
        '200':
          description: Blog translated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogTranslation'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
...